разбивает на 2 модуля заполнение большого фото
