--- js/gallery.js
@@ -1,1 +1,0 @@
-import {renderPhotos} from './thumbnails.js';
