+++ js/main.js
@@ -1,3 +1,3 @@
 import './data.js';
 import './thumbnails.js';
-import './fullpictures.js';
+import './gallary.js';
